requirements.txt
scipy==1.10.1
numpy==1.20.0
matplotlib==3.7.2
opencv_python==4.8.0.76
torch==1.2.0
torchvision==0.4.0
tqdm==4.66.1
Pillow==8.2.0
h5py==2.10.0
fightingcv_attention==1.0.0
tensorboard==2.14.0
